backend/requirements.txt
Django==5.0
djangorestframework==3.14
mysqlclient==2.2.0
openai==1.30.1
python-dotenv==1.0.1
langchain
faiss-cpu 
tiktoken